refactor(configurator): split session page modules
